+++ docs/adr/0028-react-idioms-for-overlay-components.md
@@ -21,11 +21,11 @@ Three overlay-style components do not fit a raw generated wrapper as the _public
 
 ## Decision
 
-Keep generating every wrapper with the stock `@stencil/react-output-target` (no `excludeComponents`). Hand-author a small `packages/react/src/idioms/` layer that is the public API for overlays:
+Keep generating every wrapper with the stock `@stencil/react-output-target` (no `excludeComponents`). Hand-author the public overlay API in React's conventional source groups:
 
-- **`Modal`**: wraps the generated `DsModal` and adds `onOpenChange?: (open: boolean) => void`, fired `false` on `onDsDidDismiss`, so `<Modal open={isOpen} onOpenChange={setIsOpen}>` stays in sync.
-- **`useToast()` / `useSnackbar()`**: thin hooks around `dsToastController` / `dsSnackbarController`, returning `[present, dismiss]`.
-- **`useModal()`**: scaffolded `[present, dismiss]` hook that will pass a detached element as `ModalOptions.component` once that lands in core ([#2120](https://github.com/baloise/design-system/issues/2120)).
+- **`src/components/modal.tsx`**: `Modal` wraps the generated `DsModal` and adds `onOpenChange?: (open: boolean) => void`, fired `false` on `onDsDidDismiss`, so `<Modal open={isOpen} onOpenChange={setIsOpen}>` stays in sync.
+- **`src/hooks/use-toast.ts` / `src/hooks/use-snackbar.ts`**: thin hooks around `dsToastController` / `dsSnackbarController`, returning `[present, dismiss]`.
+- **`src/hooks/use-modal.ts`**: scaffolded `[present, dismiss]` hook that will pass a detached element as `ModalOptions.component` once that lands in core ([#2120](https://github.com/baloise/design-system/issues/2120)).
 
 The generated `DsModal`, `DsToast`, `DsSnackbar`, and `DsAlertContainer`
 **values** are not re-exported from `@baloise/ds-react`. Generated overlay
@@ -39,6 +39,6 @@ This is not a custom output target. Generation stays stock; only the public barr
 
 - Overlay consumers get a React-idiomatic API instead of fighting mutable `open` or importing vanilla controllers from `@baloise/ds-core`.
 - `useToast` / `useSnackbar` must call Stencil's `defineCustomElement()` for the tags their controllers create at runtime. The shared controller hook defines `ds-alert-container` inline, while each public hook defines `ds-toast` or `ds-snackbar` inline. The generated wrappers are `/*@__PURE__*/`, so hiding them from the public barrel would otherwise let consumer bundlers drop the custom-element definitions. The type-specific imports remain separate so each hook tree-shakes independently.
-- Adding a new `ds-*` component still only requires the Stencil generator, except that `packages/react/src/components.ts` must list it if it should stay public — unit tests fail when a generated wrapper is neither hidden nor re-exported.
+- Adding a new `ds-*` component still only requires the Stencil generator, except that `packages/react/src/wrappers.ts` must list it if it should stay public — unit tests fail when a generated wrapper is neither hidden nor re-exported.
 - `useModal()` is not functional until core accepts `ModalOptions.component` (#2120). `Modal`, `useToast()`, and `useSnackbar()` do not depend on that work.
 - ADR-0003's "no `libs/output-target-react`" decision remains in force. A future Next.js/SSR wrapper would still need its own design.
