+++ packages/react/CONTEXT.md
@@ -6,18 +6,19 @@ This document captures domain language, architectural patterns, and key concepts
 
 **packages/react** (`@baloise/ds-react`) is a published wrapper around the Stencil-generated web components in `packages/core`, giving React consumers typed function-component wrappers (props, typed events, refs) instead of raw custom elements.
 
-The wrappers themselves are generated by `@stencil/react-output-target`, invoked directly from `packages/core/config/stencil.bindings.react.ts` as part of the core build (see [[packages/core|packages/core/CONTEXT.md]]). There is still no custom React output-target package — see [ADR-0003](../../docs/adr/0003-ds-react-no-custom-output-target.md). Overlay components (`ds-modal`, `ds-toast`, `ds-snackbar`) additionally get hand-authored idioms on top of those generated wrappers — see [ADR-0028](../../docs/adr/0028-react-idioms-for-overlay-components.md).
+The wrappers themselves are generated by `@stencil/react-output-target`, invoked directly from `packages/core/config/stencil.bindings.react.ts` as part of the core build (see [[packages/core|packages/core/CONTEXT.md]]). There is still no custom React output-target package — see [ADR-0003](../../docs/adr/0003-ds-react-no-custom-output-target.md). Overlay components (`ds-modal`, `ds-toast`, `ds-snackbar`) additionally get hand-authored React components and hooks on top of those generated wrappers — see [ADR-0028](../../docs/adr/0028-react-idioms-for-overlay-components.md).
 
 ## Core Concepts
 
 ### Generated vs. authored code
 
 - `src/generated/components.ts` — fully generated by the Stencil React output target on every `@baloise/ds-core` build. Never hand-edited; regenerates from scratch each build. Overlay wrappers (`DsModal`, `DsToast`, `DsSnackbar`, `DsAlertContainer`) are generated here for internal use and are **not** re-exported from the public barrel.
-- `src/components.ts` — hand-authored re-export list of every generated wrapper except the four overlay wrappers above.
-- `src/idioms/` — hand-authored React idioms that replace those overlays at the public API: controlled `Modal` (`onOpenChange`), `useToast()`, `useSnackbar()`, and a `useModal()` scaffold blocked on core `#2120`.
-- `src/index.ts` — hand-authored public barrel. Re-exports generated components via `src/components.ts`, the overlay idioms, and `bootstrapDesignSystem()`, a thin call into `@baloise/ds-core`'s `initializeDesignSystem()` for React apps that want to set brand/region/language defaults.
+- `src/wrappers.ts` — hand-authored re-export list of every generated wrapper except the four overlay wrappers above.
+- `src/components/` — hand-authored React components that replace generated wrappers at the public API, currently the controlled `Modal` (`onOpenChange`).
+- `src/hooks/` — hand-authored React hooks: `useToast()`, `useSnackbar()`, a `useModal()` scaffold blocked on core `#2120`, and their private shared helper.
+- `src/index.ts` — hand-authored public barrel. Re-exports generated components via `src/wrappers.ts`, the authored component and hooks, and `bootstrapDesignSystem()`, a thin call into `@baloise/ds-core`'s `initializeDesignSystem()` for React apps that want to set brand/region/language defaults.
 
-### Overlay idioms
+### Overlay components and hooks
 
 A **controlled overlay component** (`Modal`) wraps a generated Stencil proxy and adds `onOpenChange` so React state stays in sync when the web component self-mutates `open` (Escape, backdrop, close button).
 
@@ -29,11 +30,11 @@ The controllers create `ds-alert-container` / `ds-toast` / `ds-snackbar` with `d
 
 ### Build
 
-Plain `tsc` compile (`tsc -p tsconfig.lib.json`), matching `packages/tokens` and `libs/output-target-angular` — no bundler, since this package is a straight re-export layer plus a small idioms layer.
+Plain `tsc` compile (`tsc -p tsconfig.lib.json`), matching `packages/tokens` and `libs/output-target-angular` — no bundler, since this package is a straight re-export layer plus a small set of authored components and hooks.
 
 ### Verifying the package works
 
-Hand-authored idioms have Vitest unit tests in this package (public-hook/component behavior, with generated wrappers and core controllers mocked at the adapter boundary). End-to-end behavior (does it actually work when installed by a React app?) is verified by the integration app in `apps/integration-react/`. See `docs/adr/0009-integration-apps-as-workspace-members.md` for why that app is a normal workspace member resolving this package via `workspace:*`.
+Hand-authored components and hooks have Vitest unit tests in this package (public behavior, with generated wrappers and core controllers mocked at the adapter boundary). End-to-end behavior (does it actually work when installed by a React app?) is verified by the integration app in `apps/integration-react/`. See `docs/adr/0009-integration-apps-as-workspace-members.md` for why that app is a normal workspace member resolving this package via `workspace:*`.
 
 ## Key Constraints
